Minor fixes for Windows
- include freetds/sysdep_private.h to use correct functions;
- make DllRegisterServer/DllUnregisterServer export functions PRIVATE to
  remove a warning;
- add some casts to remove some warnings;
- disable CRT secure warnings on ODBC tests.

+++ src/odbc/native.c
@@ -201,7 +201,7 @@ parse_const_param(const char *s, TDS_SERVER_TYPE *type)
 	if (isdigit(*s) || *s == '+' || *s == '-') {
 		errno = 0;
 		strtod(s, &end);
-		if (end != s && strcspn(s, ".eE") < (end-s)&& errno == 0) {
+		if (end != s && strcspn(s, ".eE") < (size_t) (end-s) && errno == 0) {
 			*type = SYBFLT8;
 			return end;
 		}
